Added new SeperatedCollectionSize attribute
This attribute lets you link of a primitive size field/property and a
collection field/property. The size field will have its value set as the
collection's size at serialization time. Additionally the collection
will read the size field to know how many elements to read. This still
requires the use of [SendSize] on the collection though.

This feature exists for when you have cases where the size or length
information isn't adjacent in the buffer to the collection. WoW may not
do this but sometimes other protocols do.